# perf(parser-utils): adjusted siblings correction timing to reduce exponential complexity