--- a/packages/@markuplint/parser-utils/src/parser.ts
+++ b/packages/@markuplint/parser-utils/src/parser.ts
@@ -39,7 +39,7 @@ import { v4 as uuid } from 'uuid';
 
 import { attrTokenizer } from './attr-tokenizer.js';
 import { defaultSpaces } from './const.js';
-import { domLog } from './debug.js';
+import { domLog, PerformanceTimer } from './debug.js';
 import { detectElementType } from './detect-element-type.js';
 import { AttrState, TagState } from './enums.js';
 import { getEndCol, getEndLine, getOffsetsFromCode, getPosition } from './get-location.js';
@@ -47,6 +47,8 @@ import { ignoreBlock, restoreNode } from './ignore-block.js';
 import { ignoreFrontMatter } from './ignore-front-matter.js';
 import { ParserError } from './parser-error.js';
 import { sortNodes } from './sort-nodes.js';
+
+const timer = new PerformanceTimer();
 
 export abstract class Parser<Node extends {} = {}, State extends unknown = null> implements MLParser {
 	readonly #booleanish: boolean = false;
@@ -133,35 +135,53 @@ export abstract class Parser<Node extends {} = {}, State extends unknown = null>
 			// Initialize raw code
 			this.#setRawCode(rawCode, rawCode);
 
+			timer.push('beforeParse');
+			const beforeParsedCode = this.beforeParse(this.rawCode, options);
+
 			// Override raw code
-			this.#setRawCode(this.beforeParse(this.rawCode, options));
+			this.#setRawCode(beforeParsedCode);
 
 			this.#authoredElementName = options?.authoredElementName;
 
 			let frontMatter: string | null = null;
 			if (options?.ignoreFrontMatter) {
+				timer.push('ignoreFrontMatter');
 				const fm = ignoreFrontMatter(this.rawCode);
 				this.#setRawCode(fm.code);
 				frontMatter = fm.frontMatter;
 			}
 
+			timer.push('ignoreBlock');
 			const blocks = ignoreBlock(this.rawCode, this.#ignoreTags, this.#maskChar);
 			this.#setRawCode(blocks.replaced);
 
+			timer.push('tokenize');
 			const tokenized = this.tokenize(options);
 			const ast = tokenized.ast;
 			const isFragment = tokenized.isFragment;
 
 			this.#defaultDepth = options?.depth ?? this.#defaultDepth;
+
+			timer.push('traverse');
 			const traversed = this.traverse(ast, null, this.#defaultDepth);
+
+			timer.push('afterTraverse');
 			const nodeTree = this.afterTraverse([...traversed.childNodes, ...traversed.siblings]);
 
+			timer.push('flattenNodes');
 			let nodeList = this.flattenNodes(nodeTree);
+
+			timer.push('afterFlattenNodes');
 			nodeList = this.afterFlattenNodes(nodeList);
+
+			timer.push('restoreNode');
 			nodeList = restoreNode(this, nodeList, blocks, false);
+
+			timer.push('afterParse');
 			nodeList = this.afterParse(nodeList, options);
 
 			if (frontMatter) {
+				timer.push('frontMatter');
 				const newNodeList = [...nodeList];
 				let firstText = '';
 				const firstTextNode = newNodeList.shift();
@@ -186,6 +206,7 @@ export abstract class Parser<Node extends {} = {}, State extends unknown = null>
 				nodeList = [fmNode, ...newNodeList];
 			}
 
+			timer.log();
 			domLog(nodeList);
 
 			this.#reset();
@@ -232,7 +253,9 @@ export abstract class Parser<Node extends {} = {}, State extends unknown = null>
 		const siblings: MLASTNodeTreeItem[] = [];
 
 		for (const originNode of originNodes) {
+			timer.push('nodeize');
 			const nodes = this.nodeize(originNode, parentNode, depth);
+			timer.push('afterNodeize');
 			const after = this.afterNodeize(nodes, parentNode, depth);
 
 			childNodes.push(...after.siblings);
@@ -322,6 +345,7 @@ export abstract class Parser<Node extends {} = {}, State extends unknown = null>
 			readonly systemId: string;
 		},
 	): readonly MLASTNodeTreeItem[] {
+		timer.push('visitDoctype');
 		const node: MLASTDoctype = {
 			...token,
 			...this.createToken(token),
@@ -337,6 +361,7 @@ export abstract class Parser<Node extends {} = {}, State extends unknown = null>
 			readonly isBogus?: boolean;
 		},
 	): readonly MLASTNodeTreeItem[] {
+		timer.push('visitComment');
 		const isBogus = options?.isBogus ?? !token.raw.startsWith('<!--');
 		const node: MLASTComment = {
 			...token,
@@ -355,6 +380,7 @@ export abstract class Parser<Node extends {} = {}, State extends unknown = null>
 			readonly invalidTagAsText?: boolean;
 		},
 	): readonly MLASTNodeTreeItem[] {
+		timer.push('visitText');
 		const node: MLASTText = {
 			...token,
 			...this.createToken(token),
@@ -389,6 +415,7 @@ export abstract class Parser<Node extends {} = {}, State extends unknown = null>
 			readonly overwriteProps?: Partial<MLASTElement>;
 		},
 	): readonly MLASTNodeTreeItem[] {
+		timer.push('visitElement');
 		const createEndTagToken = options?.createEndTagToken;
 		const namelessFragment = options?.namelessFragment ?? false;
 		const overwriteProps = options?.overwriteProps;
@@ -453,6 +480,7 @@ export abstract class Parser<Node extends {} = {}, State extends unknown = null>
 		conditionalType: MLASTPreprocessorSpecificBlockConditionalType = null,
 		originBlockNode?: Node,
 	): readonly MLASTNodeTreeItem[] {
+		timer.push('visitPsBlock');
 		const block: MLASTPreprocessorSpecificBlock = {
 			...token,
 			...this.createToken(token),
@@ -484,6 +512,7 @@ export abstract class Parser<Node extends {} = {}, State extends unknown = null>
 	}
 
 	visitSpreadAttr(token: Token): MLASTSpreadAttr | null {
+		timer.push('visitSpreadAttr');
 		const raw = token.raw.trim();
 
 		if (raw === '') {
@@ -518,6 +547,7 @@ export abstract class Parser<Node extends {} = {}, State extends unknown = null>
 			readonly startState?: AttrState;
 		},
 	): MLASTAttr & { __rightText?: string } {
+		timer.push('visitAttr');
 		const raw = token.raw;
 
 		const quoteSet = options?.quoteSet;