ref: https://github.com/markuplint/markuplint/issues/1049

## What is the purpose of this PR?

- [ ] Fix bug
- [ ] Fix typo
- [ ] Update specs
- [ ] Add new rule
- [ ] Add new parser
- [ ] Improve or refactor rules
- [ ] Add a new core feature
- [x] Improve or refactor core features
- [ ] Update documents
- [ ] Others

### Description

N of elements | Before | After
---|---|---
1000|1300ms|300ms
2000|5000ms|650ms
3000|12000ms|2000ms
4000|23000ms|3500ms
5000|40000ms|6000ms
6000|61000ms|8900ms
7000|87000ms|13100ms